DirectConvolution

Documentation for DirectConvolution.

Use cases

These demos use data stored in the data/ folder.

data_1D = readdlm("data/signal_1.csv",',')
Plots.plot(data_1D[:,1],data_1D[:,2])

data_1D=readdlm(joinpath(dataDir,"signal_1.csv"),',')
Plots.plot(data_1D[:,1],data_1D[:,2])

Savitzky-Golay filters

Savitzky-Golay filters are a common approach to compute smoothed derivatives of a signal.

Creates a set of Savitzky-Golay filters using polynomial of degree $3$ with a window width of $11=2\times 5+1$.

sg = SG_Filter(Float64,halfWidth=5,degree=3);

SG_Filter{Float64, 11}(DirectConvolution.LinearFilter_DefaultCentered{Float64, 11}[Filter(r=-5:5,c=[-0.08392, 0.02098, 0.1026, 0.1608, 0.1958, 0.2075, 0.1958, 0.1608, 0.1026, 0.02098, -0.08392])
, Filter(r=-5:5,c=[0.05828, -0.05711, -0.1033, -0.09771, -0.0575, -2.864e-18, 0.0575, 0.09771, 0.1033, 0.05711, -0.05828])
, Filter(r=-5:5,c=[0.03497, 0.01399, -0.002331, -0.01399, -0.02098, -0.02331, -0.02098, -0.01399, -0.002331, 0.01399, 0.03497])
, Filter(r=-5:5,c=[-0.03497, 0.006993, 0.02564, 0.02681, 0.01632, 2.404e-18, -0.01632, -0.02681, -0.02564, -0.006993, 0.03497])
])

This can be checked with

julia> length(sg)11
julia> polynomialOrder(sg)3

Now apply filter:

smoothed_data_y = data_1D[:,2]

smoothed_data_y = apply_SG_filter(smoothed_data_y,sg,derivativeOrder=0)

Plots.plot(data_1D[:,1],data_1D[:,2])
Plots.plot!(data_1D[:,1],smoothed_data_y)

API

DirectConvolution.BoundaryExtension

DirectConvolution.ConstantBE

DirectConvolution.LinearFilter

DirectConvolution.LinearFilter_Default

DirectConvolution.MirrorBE

DirectConvolution.PeriodicBE

DirectConvolution.SG_Filter

DirectConvolution.UDWT_Filter

DirectConvolution.UDWT_Filter_Biorthogonal

DirectConvolution.ZeroPaddingBE

Base.filter

Base.length

Base.length

Base.range

DirectConvolution.apply_SG_filter

DirectConvolution.apply_SG_filter2D

DirectConvolution.boundaryExtension

DirectConvolution.directConv!

DirectConvolution.fcoef

DirectConvolution.maxDerivativeOrder

DirectConvolution.offset

DirectConvolution.polynomialOrder

DirectConvolution.scale

DirectConvolution.ϕ_filter

DirectConvolution.BoundaryExtension — Type

abstract type BoundaryExtension end

Abstract type associated to boundary extension.

DirectConvolution.ConstantBE — Type

struct ConstantBE <: BoundaryExtension end

A type used to tag constant constant extension

DirectConvolution.LinearFilter — Type

abstract type LinearFilter{T<:Number} end

Abstract type defining a linear filter.

DirectConvolution.LinearFilter_Default — Type

struct LinearFilter_Default{T<:Number,N}

Default linear filter.

You can create a filter as follows

julia> linear_filter=LinearFilter([1,-2,1],1)
Filter(r=-1:1,c=[1.0, -2.0, 1.0])

julia> offset(linear_filter)
1

julia> range(linear_filter)
-1:1

DirectConvolution.MirrorBE — Type

struct MirrorBE <: BoundaryExtension end

A type used to tag mirror extension

DirectConvolution.PeriodicBE — Type

struct PeriodicBE <: BoundaryExtension end

A type used to tag periodic extension

DirectConvolution.SG_Filter — Type

function SG_Filter(T::DataType=Float64;halfWidth::Int=5,degree::Int=2)

Creates a SG_Filter structure used to store Savitzky-Golay filters.

filter length is 2*halfWidth+1

polynomial degree is degree, which defines maxDerivativeOrder

You can apply these filters using the

apply_SG_filter

apply_SG_filter2D

functions.

Example:

julia> sg = SG_Filter(halfWidth=5,degree=3);

julia> maxDerivativeOrder(sg)
3

julia> length(sg)
11

julia> filter(sg,derivativeOrder=2)
Filter(r=-5:5,c=[0.03497, 0.01399, -0.002331, -0.01399, -0.02098, -0.02331, -0.02098, -0.01399, -0.002331, 0.01399, 0.03497])

DirectConvolution.UDWT_Filter — Type

abstract type UDWT_Filter{T<:Number} <: UDWT_Filter_Biorthogonal{T} end

A specialization of UDWT_Filter_Biorthogonal for orthogonal filters.

For orthogonal filters we have: ϕ=tildeϕ, ψ=tildeψ

DirectConvolution.UDWT_Filter_Biorthogonal — Type

abstract type UDWT_Filter_Biorthogonal{T<:Number} end

Abstract type defining the ϕ, ψ, tildeϕ and tildeψ filters associated to an undecimated biorthogonal wavelet transform

Subtypes of this structure are:

UDWT_Filter

Associated methods are:

ϕ_filter

ψ_filter

tildeϕ_filter

tildeψ_filter

DirectConvolution.ZeroPaddingBE — Type

struct ZeroPaddingBE <: BoundaryExtension end

A type used to tag zero padding extension

Base.filter — Method

function filter(sg::SG_Filter{T,N};derivativeOrder::Int=0)

Returns the filter to be used to compute the smoothed derivatives of order derivativeOrder.

See: SG_Filter

Base.length — Method

length(c::LinearFilter)::Int

Returns filter length

Base.length — Method

function length(sg::SG_Filter{T,N})

Returns filter length, this is an odd number.

See: SG_Filter

Base.range — Method

range(c::LinearFilter)::UnitRange

Returns filter range Ω

Filter support of a filter α is defined by Ω = [ - offset(α), length(α) - offset(α) - 1 ]

DirectConvolution.apply_SG_filter — Method

function apply_SG_filter(signal::Array{T,1},
                         sg::SG_Filter{T};
                         derivativeOrder::Int=0,
                         left_BE::Type{<:BoundaryExtension}=ConstantBE,
                         right_BE::Type{<:BoundaryExtension}=ConstantBE)

Applies an 1D Savitzky-Golay and returns the smoothed signal.

DirectConvolution.apply_SG_filter2D — Method

function apply_SG_filter2D(signal::Array{T,2},
                           sg_I::SG_Filter{T},
                           sg_J::SG_Filter{T};
                           derivativeOrder_I::Int=0,
                           derivativeOrder_J::Int=0,
                           min_I_BE::Type{<:BoundaryExtension}=ConstantBE,
                           max_I_BE::Type{<:BoundaryExtension}=ConstantBE,
                           min_J_BE::Type{<:BoundaryExtension}=ConstantBE,
                           max_J_BE::Type{<:BoundaryExtension}=ConstantBE)

Applies an 2D Savitzky-Golay and returns the smoothed signal.

DirectConvolution.boundaryExtension — Function

boundaryExtension(β::AbstractArray{T,1},
                  k::Int,
                  ::Type{BOUNDARY_EXT_TYPE})

Computes extended boundary value β[k] given boundary extension type BOUNDARY_EXT_TYPE

Available BOUNDARY_EXT_TYPE are:

ZeroPaddingBE: zero padding

ConstantBE: constant boundary extension padding

PeriodicBE: periodic boundary extension padding

MirrorBE: mirror symmetry boundary extension padding

The routine is robust in the sens that there is no restriction on k value.

julia> dom = [-5:10;];

julia> hcat(dom,map(x->boundaryExtension([1; 2; 3],x,ZeroPaddingBE),dom))'
2×16 adjoint(::Matrix{Int64}) with eltype Int64:
 -5  -4  -3  -2  -1  0  1  2  3  4  5  6  7  8  9  10
  0   0   0   0   0  0  1  2  3  0  0  0  0  0  0   0

julia> hcat(dom,map(x->boundaryExtension([1; 2; 3],x,ConstantBE),dom))'
2×16 adjoint(::Matrix{Int64}) with eltype Int64:
 -5  -4  -3  -2  -1  0  1  2  3  4  5  6  7  8  9  10
  1   1   1   1   1  1  1  2  3  3  3  3  3  3  3   3

julia> hcat(dom,map(x->boundaryExtension([1; 2; 3],x,PeriodicBE),dom))'
2×16 adjoint(::Matrix{Int64}) with eltype Int64:
 -5  -4  -3  -2  -1  0  1  2  3  4  5  6  7  8  9  10
  1   2   3   1   2  3  1  2  3  1  2  3  1  2  3   1

julia> hcat(dom,map(x->boundaryExtension([1; 2; 3],x,MirrorBE),dom))'
2×16 adjoint(::Matrix{Int64}) with eltype Int64:
 -5  -4  -3  -2  -1  0  1  2  3  4  5  6  7  8  9  10
  3   2   1   2   3  2  1  2  3  2  1  2  3  2  1   2

DirectConvolution.directConv! — Method

directConv!

Computes a convolution.

\[\gamma[k]=\sum\limits_{i\in\Omega^\alpha}\alpha[i]\beta[k+\lambda i]\]

The following example shows how to apply inplace the [0 0 1] filter on γ[5:10]

julia> β=[1:15;];

julia> γ=ones(Int,15);

julia> α=LinearFilter([0,0,1],0);

julia> directConv!(α,1,β,γ,5:10)

julia> hcat([1:length(γ);],γ)
15×2 Matrix{Int64}:
  1   1
  2   1
  3   1
  4   1
  5   7
  6   8
  7   9
  8  10
  9  11
 10  12
 11   1
 12   1
 13   1
 14   1
 15   1

DirectConvolution.fcoef — Method

fcoef(c::LinearFilter)

Returns filter coefficients

DirectConvolution.maxDerivativeOrder — Method

function maxDerivativeOrder(sg::SG_Filter{T,N})

Maximum order of the smoothed derivatives we can compute using sg filters.

See: SG_Filter

DirectConvolution.offset — Method

offset(c::LinearFilter)::Int

Returns filter offset

Caveat: the first position is 0 (and not 1)

DirectConvolution.polynomialOrder — Method

function polynomialOrder(sg::SG_Filter{T,N})

Returns the degree of the polynomial used to construct the Savitzky-Golay filters. This is mainly a 'convenience' function, as it is equivalent to maxDerivativeOrder

See: SG_Filter

DirectConvolution.scale — Method

scale(λ::Int,Ω::UnitRange{Int})

Range scaling.

Caveat:

We do not use Julia * operator as it returns a step range:

julia> r=6:8
6:8

julia> -2*r
-12:-2:-16

What we need is:

julia> r=6:8
6:8

julia> scale(-2,r)
-16:-12

DirectConvolution.ϕ_filter — Method

ϕ_filter(c::UDWT_Filter_Biorthogonal)

Returns the ϕ filter